Separate authentication and application logic
We now have an auth system with sessions. User data is encrypted with an
application secret and stored in a cookie on their browser. If this data
isn't available, we redirect to MyMLH and log the user in, redirecting
back to the original location on success. This is true of any route for
which the auth.authenticate middleware is applied.