Introduce user dashboard and improve user flow
Introduce a dashboard page where user can see all relevant links.
This will be used to remind users to complete applications, give them
a place to see application status, fill out team applications and
more. This is still not integrated with the databse to tell when
the user has completed their application or not.